# Slightly more efficient conversion from JSON to document