This PR is a small, low hanging, premature optimization.
It changes the way the `document.NewFromJSON` function works by lazily decoding the json data and delegating error handling to the `Iterate` and `GetByField` methods.
This takes advantage of the `jsonparser` library and prevents from duplicating the entire json object in memory before storing it.

It's 50% faster for small objects and about 20% for large objects, and 80% more memory efficient. It might be useful later for quick stream insertion.

```
name               old time/op    new time/op    delta
JSONToDocument-12    9.17µs ± 0%    7.86µs ± 0%   ~     (p=1.000 n=1+1)

name               old alloc/op   new alloc/op   delta
JSONToDocument-12    5.65kB ± 0%    3.10kB ± 0%   ~     (p=1.000 n=1+1)

name               old allocs/op  new allocs/op  delta
JSONToDocument-12       106 ± 0%       100 ± 0%   ~     (p=1.000 n=1+1)
```