@tie 2020-10-18
It looks like you din’t set the `-count` flag when running benchmarks (`p=1.000 n=1+1`).