@asdine 2020-10-23
10 times 👍🏼 
```
name               old time/op    new time/op    delta
JSONToDocument-12    8.39µs ± 3%    7.06µs ± 2%  -15.82%  (p=0.000 n=10+9)

name               old alloc/op   new alloc/op   delta
JSONToDocument-12    5.65kB ± 0%    3.10kB ± 0%  -45.04%  (p=0.000 n=10+10)

name               old allocs/op  new allocs/op  delta
JSONToDocument-12       106 ± 0%       100 ± 0%   -5.66%  (p=0.000 n=10+10)
```